refactor(mitm-addon): split usage buffer internals
