--- crates/runner/mitm-addon/src/usage/buffer.py
@@ -1,1193 +1,0 @@
-"""Process-local buffering for aggregate usage webhook uploads.
-
-This module owns the usage report buffer singleton used by the mitmproxy addon.
-The singleton is created on import with default settings, but its flush timer is
-scheduled lazily only after source events are accepted into the buffer.
-
-Source event idempotency keys are deduped process-wide before destination
-bucketing. The seen-key set survives flushes and is bounded by
-``MAX_SOURCE_IDEMPOTENCY_KEYS``, evicting oldest keys first, so duplicate
-response/error observations do not become separate aggregate rows.
-
-Accepted events are separated by webhook destination and output shape, then
-aggregated by ``run_id``, ``kind``, provider/model resource id, and
-``category``. Matching aggregate buckets sum ``quantity`` before delivery.
-
-Flushes are triggered by buffer bounds, the lazy timer, or explicit lifecycle
-calls. The trigger label is emitted in ``usage_event_buffer_flush`` proxy-log
-records, so callers should use the conventional labels captured by
-``UsageFlushTrigger``.
-"""
-
-from __future__ import annotations
-
-import random
-import threading
-import time
-import uuid
-from collections import OrderedDict
-from collections.abc import Callable, Iterable
-from dataclasses import dataclass, field
-from typing import Literal, Protocol, TypedDict
-
-from logging_utils import log_proxy_entry
-
-from .counters import set_buffered_usage_events
-from .idempotency import USAGE_EVENT_NAMESPACE_AGGREGATE, derive_usage_idempotency_key
-from .webhook import WebhookDeliveryOutcome, _enqueue_webhook
-
-DEFAULT_FLUSH_INTERVAL_SECONDS = 30.0
-DEFAULT_FLUSH_JITTER_RATIO = 0.2
-MAX_BUFFERED_SOURCE_EVENTS = 1_000
-MAX_AGGREGATE_BUCKETS = 100
-MAX_BUFFERED_WEBHOOK_BATCHES = 4
-MAX_SOURCE_IDEMPOTENCY_KEYS = 10_000
-USAGE_EVENT_BATCH_SIZE = 100
-MAX_RETAINED_USAGE_BATCH_RETRIES = 20
-_jitter_rng = random.SystemRandom()
-
-
-class UsageEvent(TypedDict):
-    idempotencyKey: str
-    kind: str
-    provider: str
-    category: str
-    quantity: int
-
-
-UsageFlushTrigger = Literal["timer", "threshold", "runner", "shutdown", "test"]
-ResourceFieldName = Literal["provider", "model"]
-
-
-class _TimerHandle(Protocol):
-    daemon: bool
-
-    def start(self) -> None:
-        """Start the scheduled callback."""
-
-    def cancel(self) -> None:
-        """Cancel the scheduled callback."""
-
-
-_TimerFactory = Callable[[float, Callable[[], None]], _TimerHandle]
-_DeliveryOutcomeCallback = Callable[[WebhookDeliveryOutcome], None]
-_EnqueueWebhook = Callable[[str, str, dict, str, str, _DeliveryOutcomeCallback], bool]
-
-
-class _FlushOwnerLock(Protocol):
-    def acquire(self, blocking: bool = True) -> bool:
-        raise NotImplementedError
-
-    def release(self) -> None:
-        raise NotImplementedError
-
-
-@dataclass(frozen=True)
-class _DestinationKey:
-    url: str
-    sandbox_token: str
-    proxy_log_path: str
-    resource_field_name: ResourceFieldName
-    include_kind: bool
-    log_type: str
-
-
-@dataclass(frozen=True)
-class _AggregateKey:
-    run_id: str
-    kind: str
-    provider: str
-    category: str
-
-
-@dataclass
-class _AggregateBucket:
-    quantity: int = 0
-    source_event_count: int = 0
-
-
-@dataclass(frozen=True)
-class _FlushEvent:
-    payload: dict
-    source_event_count: int
-
-
-@dataclass(frozen=True)
-class _FlushBatch:
-    url: str
-    sandbox_token: str
-    payload: dict
-    proxy_log_path: str
-    log_type: str
-    source_event_count: int
-
-
-@dataclass(frozen=True)
-class _PendingBatch:
-    batch: _FlushBatch
-    retained_retry_count: int = 0
-
-
-@dataclass
-class _FlushSummary:
-    proxy_log_path: str
-    source_event_count: int = 0
-    aggregate_event_count: int = 0
-    webhook_batch_count: int = 0
-    dropped_webhook_batch_count: int = 0
-    retained_webhook_batch_count: int = 0
-    retained_source_event_count: int = 0
-    run_ids: set[str] = field(default_factory=set)
-    destinations: set[tuple[str, str]] = field(default_factory=set)
-
-
-@dataclass
-class _PendingFlush:
-    flush_sequence: int
-    batches: list[_PendingBatch]
-    summaries: list[_FlushSummary]
-    retry_after_flush_generation: int = 0
-
-    @property
-    def source_event_count(self) -> int:
-        return sum(pending_batch.batch.source_event_count for pending_batch in self.batches)
-
-
-@dataclass
-class _DeliveringFlush:
-    pending_flush: _PendingFlush
-    trigger: UsageFlushTrigger
-    flush_generation: int
-    remaining_batch_count: int
-    retryable_batches: list[_PendingBatch] = field(default_factory=list)
-
-
-@dataclass(frozen=True)
-class _DeliveryCompletion:
-    pending_flush: _PendingFlush
-    trigger: UsageFlushTrigger
-    retained_batches: list[_PendingBatch]
-    dropped_batches: list[_PendingBatch]
-
-
-@dataclass(frozen=True)
-class _BatchAdmissionResult:
-    admitted_batch_count: int
-    retained_batches: list[_PendingBatch]
-
-
-class _BatchEnqueueError(Exception):
-    def __init__(self, original: Exception, admission_result: _BatchAdmissionResult) -> None:
-        super().__init__(str(original))
-        self.original = original
-        self.admission_result = admission_result
-
-
-@dataclass(frozen=True)
-class _RetainBatchesResult:
-    retained_batches: list[_PendingBatch]
-    dropped_batches: list[_PendingBatch]
-
-
-class _UsageBufferState:
-    """Mutable usage work state; callers must hold UsageEventBuffer._lock."""
-
-    def __init__(self) -> None:
-        self._buffer_id = str(uuid.uuid4())
-        self._flush_sequence = 0
-        self._buckets: dict[_DestinationKey, dict[_AggregateKey, _AggregateBucket]] = {}
-        # Keep source keys across flushes so aggregate idempotency does not
-        # turn response/error duplicates into distinct server-side rows.
-        self._seen_source_keys: OrderedDict[str, None] = OrderedDict()
-        self._source_event_count = 0
-        self._active_enqueue_count = 0
-        self._pending_flushes: list[_PendingFlush] = []
-        self._delivering_flushes: dict[int, _DeliveringFlush] = {}
-
-    def clear(self) -> None:
-        self._buckets = {}
-        self._seen_source_keys.clear()
-        self._source_event_count = 0
-        self._active_enqueue_count = 0
-        self._pending_flushes = []
-        self._delivering_flushes = {}
-
-    def add_events(
-        self,
-        url: str,
-        sandbox_token: str,
-        run_id: str,
-        events: Iterable[UsageEvent],
-        proxy_log_path: str,
-        *,
-        resource_field_name: ResourceFieldName,
-        include_kind: bool,
-        log_type: str,
-    ) -> int:
-        buckets: dict[_AggregateKey, _AggregateBucket] | None = None
-        destination = _DestinationKey(
-            url,
-            sandbox_token,
-            proxy_log_path,
-            resource_field_name,
-            include_kind,
-            log_type,
-        )
-        accepted_count = 0
-        for event in events:
-            source_key = event["idempotencyKey"]
-            if source_key in self._seen_source_keys:
-                continue
-            if buckets is None:
-                buckets = self._buckets.setdefault(destination, {})
-            self._seen_source_keys[source_key] = None
-            aggregate_key = _AggregateKey(
-                run_id=run_id,
-                kind=event["kind"],
-                provider=event["provider"],
-                category=event["category"],
-            )
-            bucket = buckets.setdefault(aggregate_key, _AggregateBucket())
-            bucket.quantity += event["quantity"]
-            bucket.source_event_count += 1
-            self._source_event_count += 1
-            accepted_count += 1
-        self._evict_source_keys()
-        return accepted_count
-
-    def _evict_source_keys(self) -> None:
-        while len(self._seen_source_keys) > MAX_SOURCE_IDEMPOTENCY_KEYS:
-            self._seen_source_keys.popitem(last=False)
-
-    def should_flush(self) -> bool:
-        if self._source_event_count >= MAX_BUFFERED_SOURCE_EVENTS:
-            return True
-        if sum(len(buckets) for buckets in self._buckets.values()) >= MAX_AGGREGATE_BUCKETS:
-            return True
-        return self._estimated_webhook_batch_count() >= MAX_BUFFERED_WEBHOOK_BATCHES
-
-    def _estimated_webhook_batch_count(self) -> int:
-        count = 0
-        for buckets in self._buckets.values():
-            events_by_run: dict[str, int] = {}
-            for aggregate_key in buckets:
-                events_by_run[aggregate_key.run_id] = events_by_run.get(aggregate_key.run_id, 0) + 1
-            count += sum(
-                (event_count + USAGE_EVENT_BATCH_SIZE - 1) // USAGE_EVENT_BATCH_SIZE
-                for event_count in events_by_run.values()
-            )
-        return count
-
-    def has_schedulable_work(self) -> bool:
-        return bool(self._pending_flushes or self._source_event_count)
-
-    def has_active_enqueue(self) -> bool:
-        return bool(self._active_enqueue_count)
-
-    def has_available_pending_flushes(self, flush_generation: int) -> bool:
-        return any(
-            pending_flush.retry_after_flush_generation != flush_generation
-            for pending_flush in self._pending_flushes
-        )
-
-    def pending_flush_priority(self, flush_generation: int) -> int | None:
-        priorities = [
-            _pending_flush_priority(pending_flush)
-            for pending_flush in self._pending_flushes
-            if pending_flush.retry_after_flush_generation != flush_generation
-        ]
-        if not priorities:
-            return None
-        return min(priorities)
-
-    def pop_highest_priority_pending_flush(self, flush_generation: int) -> _PendingFlush:
-        selected_index: int | None = None
-        selected_priority: int | None = None
-        for index, pending_flush in enumerate(self._pending_flushes):
-            if pending_flush.retry_after_flush_generation == flush_generation:
-                continue
-            priority = _pending_flush_priority(pending_flush)
-            if selected_priority is None or priority < selected_priority:
-                selected_index = index
-                selected_priority = priority
-        if selected_index is None:
-            raise RuntimeError("no pending flush is available for this flush generation")
-        pending_flush = self._pending_flushes.pop(selected_index)
-        pending_flush.retry_after_flush_generation = 0
-        return pending_flush
-
-    def live_priority(self) -> int | None:
-        if not self._buckets:
-            return None
-        return min(_destination_priority(destination) for destination in self._buckets)
-
-    def snapshot_live_flush(self) -> _PendingFlush | None:
-        if not self._buckets:
-            self._source_event_count = 0
-            return None
-
-        self._flush_sequence += 1
-        flush_sequence = self._flush_sequence
-        buckets = self._buckets
-        self._buckets = {}
-        self._source_event_count = 0
-        batches = self._build_flush_batches(buckets, flush_sequence)
-        return _pending_flush_from_batches(flush_sequence, batches)
-
-    def begin_delivery(
-        self,
-        pending_flush: _PendingFlush,
-        trigger: UsageFlushTrigger,
-        flush_generation: int,
-    ) -> None:
-        self._active_enqueue_count += 1
-        self._delivering_flushes[id(pending_flush)] = _DeliveringFlush(
-            pending_flush=pending_flush,
-            trigger=trigger,
-            flush_generation=flush_generation,
-            remaining_batch_count=len(pending_flush.batches),
-        )
-
-    def complete_enqueue(self) -> None:
-        self._active_enqueue_count = max(0, self._active_enqueue_count - 1)
-
-    def fail_enqueue(
-        self, pending_flush: _PendingFlush, flush_generation: int
-    ) -> _RetainBatchesResult:
-        if self._delivering_flushes.pop(id(pending_flush), None) is not None:
-            retain_result = self._retain_batches_for_retry(
-                pending_flush.flush_sequence,
-                pending_flush.batches,
-                flush_generation,
-            )
-        else:
-            retain_result = _RetainBatchesResult(retained_batches=[], dropped_batches=[])
-        self.complete_enqueue()
-        return retain_result
-
-    def record_delivery_outcome(
-        self,
-        pending_flush: _PendingFlush,
-        pending_batch: _PendingBatch,
-        outcome: WebhookDeliveryOutcome,
-    ) -> _DeliveryCompletion | None:
-        delivering_flush = self._delivering_flushes.get(id(pending_flush))
-        if delivering_flush is None:
-            return None
-
-        if outcome == "retryable_failure":
-            delivering_flush.retryable_batches.append(pending_batch)
-        delivering_flush.remaining_batch_count -= 1
-        if delivering_flush.remaining_batch_count > 0:
-            return None
-
-        del self._delivering_flushes[id(pending_flush)]
-        completed_flush = delivering_flush.pending_flush
-        retryable_batch_ids = {
-            id(pending_batch) for pending_batch in delivering_flush.retryable_batches
-        }
-        retryable_batches = [
-            pending_batch
-            for pending_batch in completed_flush.batches
-            if id(pending_batch) in retryable_batch_ids
-        ]
-        retain_result = self._retain_batches_for_retry(
-            completed_flush.flush_sequence,
-            retryable_batches,
-            delivering_flush.flush_generation,
-        )
-        return _DeliveryCompletion(
-            pending_flush=completed_flush,
-            trigger=delivering_flush.trigger,
-            retained_batches=retain_result.retained_batches,
-            dropped_batches=retain_result.dropped_batches,
-        )
-
-    def complete_admission(
-        self, pending_flush: _PendingFlush, admission_result: _BatchAdmissionResult
-    ) -> _RetainBatchesResult:
-        delivering_flush = self._delivering_flushes.get(id(pending_flush))
-        retain_result = _RetainBatchesResult(retained_batches=[], dropped_batches=[])
-        if delivering_flush is not None and admission_result.retained_batches:
-            admitted_batches = pending_flush.batches[: admission_result.admitted_batch_count]
-            completed_outcome_count = (
-                len(pending_flush.batches) - delivering_flush.remaining_batch_count
-            )
-            remaining_admitted_count = (
-                admission_result.admitted_batch_count - completed_outcome_count
-            )
-            if admitted_batches and remaining_admitted_count > 0:
-                retain_result = self._retain_batches_for_retry(
-                    pending_flush.flush_sequence,
-                    admission_result.retained_batches,
-                    delivering_flush.flush_generation,
-                )
-                delivering_flush.pending_flush = _pending_flush_from_pending_batches(
-                    pending_flush.flush_sequence,
-                    admitted_batches,
-                )
-                delivering_flush.remaining_batch_count = remaining_admitted_count
-            else:
-                retryable_batch_ids = {
-                    id(pending_batch) for pending_batch in delivering_flush.retryable_batches
-                }
-                batches_to_retain = [
-                    pending_batch
-                    for index, pending_batch in enumerate(pending_flush.batches)
-                    if index >= admission_result.admitted_batch_count
-                    or id(pending_batch) in retryable_batch_ids
-                ]
-                retain_result = self._retain_batches_for_retry(
-                    pending_flush.flush_sequence,
-                    batches_to_retain,
-                    delivering_flush.flush_generation,
-                )
-                del self._delivering_flushes[id(pending_flush)]
-        self.complete_enqueue()
-        return retain_result
-
-    def _retain_batches_for_retry(
-        self,
-        flush_sequence: int,
-        pending_batches: list[_PendingBatch],
-        flush_generation: int,
-    ) -> _RetainBatchesResult:
-        retained_batches: list[_PendingBatch] = []
-        dropped_batches: list[_PendingBatch] = []
-        for pending_batch in pending_batches:
-            if pending_batch.retained_retry_count >= MAX_RETAINED_USAGE_BATCH_RETRIES:
-                dropped_batches.append(pending_batch)
-            else:
-                retained_batches.append(
-                    _PendingBatch(
-                        batch=pending_batch.batch,
-                        retained_retry_count=pending_batch.retained_retry_count + 1,
-                    )
-                )
-        if retained_batches:
-            retained_flush = _pending_flush_from_pending_batches(flush_sequence, retained_batches)
-            retained_flush.retry_after_flush_generation = flush_generation
-            self._pending_flushes.append(retained_flush)
-        return _RetainBatchesResult(
-            retained_batches=retained_batches,
-            dropped_batches=dropped_batches,
-        )
-
-    def buffered_source_event_count(self) -> int:
-        return (
-            self._source_event_count
-            + sum(pending_flush.source_event_count for pending_flush in self._pending_flushes)
-            + sum(
-                delivering_flush.pending_flush.source_event_count
-                for delivering_flush in self._delivering_flushes.values()
-            )
-        )
-
-    def _build_flush_batches(
-        self,
-        buckets: dict[_DestinationKey, dict[_AggregateKey, _AggregateBucket]],
-        flush_sequence: int,
-    ) -> list[_FlushBatch]:
-        batches: list[_FlushBatch] = []
-        for destination in sorted(
-            buckets,
-            key=lambda item: (
-                _destination_priority(item),
-                item.url,
-                item.sandbox_token,
-                item.proxy_log_path,
-                item.resource_field_name,
-                item.include_kind,
-                item.log_type,
-            ),
-        ):
-            events_by_run = self._events_by_run(destination, buckets[destination], flush_sequence)
-            for run_id in sorted(events_by_run):
-                events = events_by_run[run_id]
-                for start in range(0, len(events), USAGE_EVENT_BATCH_SIZE):
-                    batch_events = events[start : start + USAGE_EVENT_BATCH_SIZE]
-                    batches.append(
-                        _FlushBatch(
-                            url=destination.url,
-                            sandbox_token=destination.sandbox_token,
-                            payload={
-                                "runId": run_id,
-                                "events": [event.payload for event in batch_events],
-                            },
-                            proxy_log_path=destination.proxy_log_path,
-                            log_type=destination.log_type,
-                            source_event_count=sum(
-                                event.source_event_count for event in batch_events
-                            ),
-                        )
-                    )
-        return batches
-
-    def _events_by_run(
-        self,
-        destination: _DestinationKey,
-        buckets: dict[_AggregateKey, _AggregateBucket],
-        flush_sequence: int,
-    ) -> dict[str, list[_FlushEvent]]:
-        events_by_run: dict[str, list[_FlushEvent]] = {}
-        for aggregate_key in sorted(
-            buckets,
-            key=lambda item: (item.run_id, item.kind, item.provider, item.category),
-        ):
-            bucket = buckets[aggregate_key]
-            event = _FlushEvent(
-                payload={
-                    "idempotencyKey": self._aggregate_idempotency_key(
-                        destination, aggregate_key, flush_sequence
-                    ),
-                    destination.resource_field_name: aggregate_key.provider,
-                    "category": aggregate_key.category,
-                    "quantity": bucket.quantity,
-                },
-                source_event_count=bucket.source_event_count,
-            )
-            if destination.include_kind:
-                event.payload["kind"] = aggregate_key.kind
-            events_by_run.setdefault(aggregate_key.run_id, []).append(event)
-        return events_by_run
-
-    def _aggregate_idempotency_key(
-        self,
-        destination: _DestinationKey,
-        aggregate_key: _AggregateKey,
-        flush_sequence: int,
-    ) -> str:
-        return derive_usage_idempotency_key(
-            USAGE_EVENT_NAMESPACE_AGGREGATE,
-            (
-                self._buffer_id,
-                str(flush_sequence),
-                destination.url,
-                destination.sandbox_token,
-                destination.proxy_log_path,
-                aggregate_key.run_id,
-                aggregate_key.kind,
-                aggregate_key.provider,
-                aggregate_key.category,
-            ),
-        )
-
-
-class UsageEventBuffer:
-    """Thread-safe process-local usage report buffer."""
-
-    def __init__(
-        self,
-        *,
-        flush_interval_seconds: float = DEFAULT_FLUSH_INTERVAL_SECONDS,
-        jitter_ratio: float = DEFAULT_FLUSH_JITTER_RATIO,
-        timer_enabled: bool = True,
-        timer_factory: _TimerFactory | None = None,
-        enqueue_webhook: _EnqueueWebhook | None = None,
-        flush_owner_lock: _FlushOwnerLock | None = None,
-    ) -> None:
-        self._lock = threading.Lock()
-        # Serializes snapshot/enqueue ownership. Ordinary flushes defer if busy;
-        # lifecycle drains wait so their acknowledgement snapshots follow drain
-        # ownership instead of only recording signal receipt.
-        self._flush_owner_lock: _FlushOwnerLock = (
-            flush_owner_lock if flush_owner_lock is not None else threading.Lock()
-        )
-        self._enqueue_webhook = enqueue_webhook
-        self._state = _UsageBufferState()
-        self._flush_interval_seconds = max(1.0, flush_interval_seconds)
-        self._jitter_ratio = max(0.0, jitter_ratio)
-        self._timer_enabled = timer_enabled
-        self._timer_factory = timer_factory if timer_factory is not None else self._make_timer
-        self._timer: _TimerHandle | None = None
-        self._flush_generation = 0
-
-    def configure(self, *, flush_interval_seconds: float) -> None:
-        """Update runtime buffer settings."""
-        with self._lock:
-            self._flush_interval_seconds = max(1.0, flush_interval_seconds)
-
-    def buffer_usage_events(
-        self,
-        url: str,
-        sandbox_token: str,
-        run_id: str,
-        events: Iterable[UsageEvent],
-        proxy_log_path: str,
-        *,
-        resource_field_name: ResourceFieldName = "provider",
-        include_kind: bool = True,
-        log_type: str = "usage_event",
-    ) -> int:
-        """Add source usage events and flush if the buffer exceeds a bound."""
-        flush_now = False
-        timer_to_start: _TimerHandle | None = None
-        with self._lock:
-            accepted_count = self._state.add_events(
-                url,
-                sandbox_token,
-                run_id,
-                events,
-                proxy_log_path,
-                resource_field_name=resource_field_name,
-                include_kind=include_kind,
-                log_type=log_type,
-            )
-            if accepted_count == 0:
-                return 0
-            if self._state.should_flush():
-                flush_now = True
-            else:
-                timer_to_start = self._schedule_timer_locked()
-            self._sync_buffered_counter_locked()
-
-        if timer_to_start is not None:
-            timer_to_start.start()
-        if flush_now:
-            self._flush_usage_events(trigger="threshold")
-        return accepted_count
-
-    def flush_usage_events(self, *, trigger: UsageFlushTrigger) -> int:
-        """Flush all buffered usage events now."""
-        return self._flush_usage_events(trigger=trigger)
-
-    def close(self) -> None:
-        """Cancel any pending timer for test cleanup or process shutdown."""
-        with self._lock:
-            timer = self._pop_timer_locked()
-            self._state.clear()
-            self._sync_buffered_counter_locked()
-        if timer is not None:
-            timer.cancel()
-
-    def _flush_usage_events(self, *, trigger: UsageFlushTrigger) -> int:
-        if not self._acquire_flush_ownership(trigger):
-            self._defer_unowned_flush(trigger)
-            return 0
-
-        try:
-            return self._flush_usage_events_owned(trigger=trigger)
-        finally:
-            self._flush_owner_lock.release()
-
-    def _acquire_flush_ownership(self, trigger: UsageFlushTrigger) -> bool:
-        return self._flush_owner_lock.acquire(blocking=trigger in ("runner", "shutdown"))
-
-    def _defer_unowned_flush(self, trigger: UsageFlushTrigger) -> None:
-        timer_to_cancel: _TimerHandle | None = None
-        timer_to_start: _TimerHandle | None = None
-        with self._lock:
-            if trigger == "timer":
-                timer_to_cancel = self._pop_timer_locked()
-            if trigger != "shutdown":
-                timer_to_start = self._schedule_timer_if_buffered_locked()
-            self._sync_buffered_counter_locked()
-        if timer_to_cancel is not None:
-            timer_to_cancel.cancel()
-        if timer_to_start is not None:
-            timer_to_start.start()
-
-    def _flush_usage_events_owned(self, *, trigger: UsageFlushTrigger) -> int:
-        flushed_batch_count = 0
-        snapshot_live = True
-        self._flush_generation += 1
-        flush_generation = self._flush_generation
-        if trigger in ("shutdown", "timer"):
-            with self._lock:
-                timer = self._pop_timer_locked()
-            if timer is not None:
-                timer.cancel()
-
-        while True:
-            timer_to_start: _TimerHandle | None = None
-            with self._lock:
-                pending_flush, live_snapshot_attempted = self._next_pending_flush_locked(
-                    snapshot_live=snapshot_live,
-                    flush_generation=flush_generation,
-                )
-                if live_snapshot_attempted and trigger != "shutdown":
-                    snapshot_live = False
-                if pending_flush is None:
-                    if trigger != "shutdown":
-                        timer_to_start = self._schedule_timer_if_buffered_locked()
-                    self._sync_buffered_counter_locked()
-                else:
-                    self._state.begin_delivery(pending_flush, trigger, flush_generation)
-                    self._sync_buffered_counter_locked()
-
-            if timer_to_start is not None:
-                timer_to_start.start()
-            if pending_flush is None:
-                return flushed_batch_count
-
-            try:
-                admission_result = self._enqueue_pending_flush(pending_flush, trigger)
-            except _BatchEnqueueError as exc:
-                timer_to_start = None
-                with self._lock:
-                    retain_result = self._state.complete_admission(
-                        pending_flush, exc.admission_result
-                    )
-                    if retain_result.retained_batches and trigger != "shutdown":
-                        timer_to_start = self._schedule_timer_if_buffered_locked()
-                    self._sync_buffered_counter_locked()
-                if retain_result.dropped_batches:
-                    _log_dropped_batches(
-                        trigger,
-                        pending_flush.flush_sequence,
-                        retain_result.dropped_batches,
-                    )
-                if timer_to_start is not None:
-                    timer_to_start.start()
-                raise exc.original from exc
-            except Exception:
-                timer_to_start = None
-                with self._lock:
-                    retain_result = self._state.fail_enqueue(pending_flush, flush_generation)
-                    if trigger != "shutdown":
-                        timer_to_start = self._schedule_timer_if_buffered_locked()
-                    self._sync_buffered_counter_locked()
-                if retain_result.dropped_batches:
-                    _log_dropped_batches(
-                        trigger,
-                        pending_flush.flush_sequence,
-                        retain_result.dropped_batches,
-                    )
-                if timer_to_start is not None:
-                    timer_to_start.start()
-                raise
-
-            flushed_batch_count += admission_result.admitted_batch_count
-            timer_to_start = None
-            with self._lock:
-                retain_result = self._state.complete_admission(pending_flush, admission_result)
-                if retain_result.retained_batches and trigger != "shutdown":
-                    timer_to_start = self._schedule_timer_if_buffered_locked()
-                self._sync_buffered_counter_locked()
-            if retain_result.dropped_batches:
-                _log_dropped_batches(
-                    trigger,
-                    pending_flush.flush_sequence,
-                    retain_result.dropped_batches,
-                )
-            if timer_to_start is not None:
-                timer_to_start.start()
-            if admission_result.retained_batches:
-                return flushed_batch_count
-
-    def _enqueue_pending_flush(
-        self,
-        pending_flush: _PendingFlush,
-        trigger: UsageFlushTrigger,
-    ) -> _BatchAdmissionResult:
-        started_at = time.monotonic()
-        try:
-            _log_flush_summaries(
-                "started", trigger, pending_flush.flush_sequence, pending_flush.summaries
-            )
-            admission_result = _enqueue_batches(
-                pending_flush.batches,
-                self._enqueue_webhook if self._enqueue_webhook is not None else _enqueue_webhook,
-                lambda pending_batch: self._make_delivery_outcome_callback(
-                    pending_flush, pending_batch
-                ),
-            )
-            _apply_retained_batch_counts(pending_flush.summaries, admission_result.retained_batches)
-            _log_flush_summaries(
-                "enqueued",
-                trigger,
-                pending_flush.flush_sequence,
-                pending_flush.summaries,
-                duration_ms=_elapsed_ms(started_at),
-            )
-            return admission_result
-        except Exception as exc:
-            error_type = (
-                type(exc.original).__name__
-                if isinstance(exc, _BatchEnqueueError)
-                else type(exc).__name__
-            )
-            _log_flush_summaries(
-                "failed",
-                trigger,
-                pending_flush.flush_sequence,
-                pending_flush.summaries,
-                duration_ms=_elapsed_ms(started_at),
-                error_type=error_type,
-            )
-            raise
-
-    def _make_delivery_outcome_callback(
-        self,
-        pending_flush: _PendingFlush,
-        pending_batch: _PendingBatch,
-    ) -> _DeliveryOutcomeCallback:
-        def callback(outcome: WebhookDeliveryOutcome) -> None:
-            self._record_delivery_outcome(pending_flush, pending_batch, outcome)
-
-        return callback
-
-    def _record_delivery_outcome(
-        self,
-        pending_flush: _PendingFlush,
-        pending_batch: _PendingBatch,
-        outcome: WebhookDeliveryOutcome,
-    ) -> None:
-        timer_to_start: _TimerHandle | None = None
-        completion: _DeliveryCompletion | None = None
-        with self._lock:
-            completion = self._state.record_delivery_outcome(pending_flush, pending_batch, outcome)
-            if (
-                completion is not None
-                and completion.retained_batches
-                and completion.trigger != "shutdown"
-            ):
-                timer_to_start = self._schedule_timer_if_buffered_locked()
-            self._sync_buffered_counter_locked()
-
-        if completion is not None and completion.retained_batches:
-            _log_flush_summaries(
-                "retained",
-                completion.trigger,
-                completion.pending_flush.flush_sequence,
-                _build_flush_summaries(
-                    [pending_batch.batch for pending_batch in completion.retained_batches]
-                ),
-            )
-        if completion is not None and completion.dropped_batches:
-            _log_dropped_batches(
-                completion.trigger,
-                completion.pending_flush.flush_sequence,
-                completion.dropped_batches,
-            )
-        if timer_to_start is not None:
-            timer_to_start.start()
-
-    def _sync_buffered_counter_locked(self) -> None:
-        set_buffered_usage_events(self._state.buffered_source_event_count())
-
-    def _schedule_timer_locked(self) -> _TimerHandle | None:
-        if not self._timer_enabled or self._timer is not None:
-            return None
-        delay = self._next_delay_seconds()
-        timer = self._timer_factory(delay, self._flush_from_timer)
-        timer.daemon = True
-        self._timer = timer
-        return timer
-
-    def _schedule_timer_if_buffered_locked(self) -> _TimerHandle | None:
-        if not self._state.has_schedulable_work():
-            return None
-        return self._schedule_timer_locked()
-
-    def _pop_timer_locked(self) -> _TimerHandle | None:
-        timer = self._timer
-        self._timer = None
-        return timer
-
-    def _flush_from_timer(self) -> None:
-        self.flush_usage_events(trigger="timer")
-
-    def _next_delay_seconds(self) -> float:
-        jitter = self._flush_interval_seconds * self._jitter_ratio
-        return max(0.001, self._flush_interval_seconds + _jitter_rng.uniform(-jitter, jitter))
-
-    def _next_pending_flush_locked(
-        self, *, snapshot_live: bool, flush_generation: int
-    ) -> tuple[_PendingFlush | None, bool]:
-        if self._state.has_active_enqueue():
-            return None, False
-        if self._state.has_available_pending_flushes(flush_generation):
-            timer = self._pop_timer_locked()
-            if timer is not None:
-                timer.cancel()
-            pending_priority = self._state.pending_flush_priority(flush_generation)
-            live_priority = self._state.live_priority() if snapshot_live else None
-            if (
-                pending_priority is not None
-                and live_priority is not None
-                and live_priority < pending_priority
-            ):
-                return self._state.snapshot_live_flush(), True
-            return self._state.pop_highest_priority_pending_flush(flush_generation), False
-        if not snapshot_live:
-            return None, False
-        timer = self._pop_timer_locked()
-        if timer is not None:
-            timer.cancel()
-        return self._state.snapshot_live_flush(), True
-
-    @staticmethod
-    def _make_timer(delay: float, callback: Callable[[], None]) -> threading.Timer:
-        return threading.Timer(delay, callback)
-
-
-def _enqueue_batches(
-    batches: list[_PendingBatch],
-    enqueue_webhook: _EnqueueWebhook,
-    delivery_outcome_callback: Callable[[_PendingBatch], _DeliveryOutcomeCallback],
-) -> _BatchAdmissionResult:
-    for index, pending_batch in enumerate(batches):
-        batch = pending_batch.batch
-        try:
-            admitted = enqueue_webhook(
-                batch.url,
-                batch.sandbox_token,
-                batch.payload,
-                batch.proxy_log_path,
-                batch.log_type,
-                delivery_outcome_callback(pending_batch),
-            )
-        except Exception as exc:
-            raise _BatchEnqueueError(
-                exc,
-                _BatchAdmissionResult(
-                    admitted_batch_count=index,
-                    retained_batches=batches[index:],
-                ),
-            ) from exc
-        if admitted is False:
-            return _BatchAdmissionResult(
-                admitted_batch_count=index,
-                retained_batches=batches[index:],
-            )
-    return _BatchAdmissionResult(
-        admitted_batch_count=len(batches),
-        retained_batches=[],
-    )
-
-
-def _apply_retained_batch_counts(
-    summaries: Iterable[_FlushSummary],
-    retained_batches: Iterable[_PendingBatch],
-) -> None:
-    retained_batch_counts: dict[str, int] = {}
-    retained_source_counts: dict[str, int] = {}
-    for pending_batch in retained_batches:
-        batch = pending_batch.batch
-        retained_batch_counts[batch.proxy_log_path] = (
-            retained_batch_counts.get(batch.proxy_log_path, 0) + 1
-        )
-        retained_source_counts[batch.proxy_log_path] = (
-            retained_source_counts.get(batch.proxy_log_path, 0) + batch.source_event_count
-        )
-    for summary in summaries:
-        summary.retained_webhook_batch_count = retained_batch_counts.get(summary.proxy_log_path, 0)
-        summary.retained_source_event_count = retained_source_counts.get(summary.proxy_log_path, 0)
-
-
-def _build_flush_summaries(batches: Iterable[_FlushBatch]) -> list[_FlushSummary]:
-    summaries: dict[str, _FlushSummary] = {}
-    for batch in batches:
-        summary = summaries.setdefault(
-            batch.proxy_log_path,
-            _FlushSummary(proxy_log_path=batch.proxy_log_path),
-        )
-        summary.source_event_count += batch.source_event_count
-        events = batch.payload.get("events")
-        if isinstance(events, list):
-            summary.aggregate_event_count += len(events)
-        summary.webhook_batch_count += 1
-        run_id = batch.payload.get("runId")
-        if isinstance(run_id, str) and run_id:
-            summary.run_ids.add(run_id)
-        summary.destinations.add((batch.url, batch.proxy_log_path))
-
-    return [summaries[path] for path in sorted(summaries)]
-
-
-def _pending_flush_from_batches(flush_sequence: int, batches: list[_FlushBatch]) -> _PendingFlush:
-    return _pending_flush_from_pending_batches(
-        flush_sequence,
-        [_PendingBatch(batch=batch) for batch in batches],
-    )
-
-
-def _pending_flush_from_pending_batches(
-    flush_sequence: int, batches: list[_PendingBatch]
-) -> _PendingFlush:
-    return _PendingFlush(
-        flush_sequence=flush_sequence,
-        batches=batches,
-        summaries=_build_flush_summaries([pending_batch.batch for pending_batch in batches]),
-    )
-
-
-def _destination_priority(destination: _DestinationKey) -> int:
-    return _log_type_priority(destination.log_type)
-
-
-def _pending_flush_priority(pending_flush: _PendingFlush) -> int:
-    if not pending_flush.batches:
-        return 1
-    return min(_batch_priority(pending_batch.batch) for pending_batch in pending_flush.batches)
-
-
-def _batch_priority(batch: _FlushBatch) -> int:
-    return _log_type_priority(batch.log_type)
-
-
-def _log_type_priority(log_type: str) -> int:
-    if log_type == "usage_event":
-        return 0
-    return 1
-
-
-def _log_dropped_batches(
-    trigger: UsageFlushTrigger,
-    flush_sequence: int,
-    dropped_batches: list[_PendingBatch],
-) -> None:
-    if not dropped_batches:
-        return
-    _log_flush_summaries(
-        "dropped",
-        trigger,
-        flush_sequence,
-        _build_flush_summaries([pending_batch.batch for pending_batch in dropped_batches]),
-        reason="retry_budget_exhausted",
-        retained_retry_count=max(
-            pending_batch.retained_retry_count for pending_batch in dropped_batches
-        ),
-    )
-
-
-def _log_flush_summaries(
-    phase: str,
-    trigger: UsageFlushTrigger,
-    flush_sequence: int,
-    summaries: Iterable[_FlushSummary],
-    *,
-    duration_ms: int | None = None,
-    error_type: str | None = None,
-    reason: str | None = None,
-    retained_retry_count: int | None = None,
-) -> None:
-    for summary in summaries:
-        if not summary.proxy_log_path:
-            continue
-        retained_webhook_batch_count = summary.retained_webhook_batch_count
-        retained_source_event_count = summary.retained_source_event_count
-        if phase == "retained":
-            retained_webhook_batch_count = summary.webhook_batch_count
-            retained_source_event_count = summary.source_event_count
-        extra: dict[str, object] = {
-            "type": "usage_event_buffer_flush",
-            "phase": phase,
-            "trigger": trigger,
-            "flush_sequence": flush_sequence,
-            "source_event_count": summary.source_event_count,
-            "aggregate_event_count": summary.aggregate_event_count,
-            "webhook_batch_count": summary.webhook_batch_count,
-            "dropped_webhook_batch_count": summary.dropped_webhook_batch_count,
-            "retained_webhook_batch_count": retained_webhook_batch_count,
-            "retained_source_event_count": retained_source_event_count,
-            "run_count": len(summary.run_ids),
-            "destination_count": len(summary.destinations),
-        }
-        if phase == "dropped":
-            extra["dropped_webhook_batch_count"] = summary.webhook_batch_count
-            extra["dropped_source_event_count"] = summary.source_event_count
-        if duration_ms is not None:
-            extra["duration_ms"] = duration_ms
-        if error_type is not None:
-            extra["error_type"] = error_type
-        if reason is not None:
-            extra["reason"] = reason
-        if retained_retry_count is not None:
-            extra["retained_retry_count"] = retained_retry_count
-        level = "error" if phase == "failed" else "info"
-        message = f"Usage event buffer flush {phase}"
-        if phase == "retained":
-            message = "Usage event buffer flush retained for retry"
-        elif phase == "dropped":
-            level = "error"
-            message = "Usage event buffer flush dropped retained batches"
-        elif phase == "enqueued" and summary.retained_webhook_batch_count:
-            message = "Usage event buffer flush retained webhook batches for retry"
-        log_proxy_entry(
-            summary.proxy_log_path,
-            level,
-            message,
-            **extra,
-        )
-
-
-def _elapsed_ms(started_at: float) -> int:
-    return max(0, int((time.monotonic() - started_at) * 1000))
-
-
-_usage_event_buffer = UsageEventBuffer()
-
-
-def configure_usage_buffer(*, flush_interval_seconds: float) -> None:
-    """Update singleton buffer settings for future timer scheduling.
-
-    Existing scheduled timers are not rescheduled.
-    """
-    _usage_event_buffer.configure(flush_interval_seconds=flush_interval_seconds)
-
-
-def buffer_usage_events(
-    url: str,
-    sandbox_token: str,
-    run_id: str,
-    events: Iterable[UsageEvent],
-    proxy_log_path: str,
-) -> int:
-    """Buffer source events on the singleton and return the accepted count.
-
-    Source idempotency-key duplicates are dropped before aggregation, so the
-    accepted count can be smaller than the number of input events. A threshold
-    flush may be enqueued before this returns.
-    """
-    return _usage_event_buffer.buffer_usage_events(
-        url,
-        sandbox_token,
-        run_id,
-        events,
-        proxy_log_path,
-    )
-
-
-def buffer_model_usage_observations(
-    url: str,
-    sandbox_token: str,
-    run_id: str,
-    events: Iterable[UsageEvent],
-    proxy_log_path: str,
-) -> int:
-    """Buffer model usage observations with the observation webhook shape."""
-    return _usage_event_buffer.buffer_usage_events(
-        url,
-        sandbox_token,
-        run_id,
-        events,
-        proxy_log_path,
-        resource_field_name="model",
-        include_kind=False,
-        log_type="model_usage_observation",
-    )
-
-
-def flush_usage_events(*, trigger: UsageFlushTrigger) -> int:
-    """Flush the singleton, log the trigger, and return the webhook batch count."""
-    return _usage_event_buffer.flush_usage_events(trigger=trigger)
-
-
-def reset_usage_buffer_for_tests(
-    *,
-    timer_enabled: bool = False,
-    timer_factory: _TimerFactory | None = None,
-    enqueue_webhook: _EnqueueWebhook | None = None,
-    flush_owner_lock: _FlushOwnerLock | None = None,
-) -> None:
-    """Cancel pending timer work and replace singleton state for test isolation."""
-    global _usage_event_buffer
-    _usage_event_buffer.close()
-    _usage_event_buffer = UsageEventBuffer(
-        timer_enabled=timer_enabled,
-        timer_factory=timer_factory,
-        enqueue_webhook=enqueue_webhook,
-        flush_owner_lock=flush_owner_lock,
-    )
